+++ deepparse/parser/address_parser.py
@@ -6,11 +6,11 @@
 
 import contextlib
 import os
-import platform
 import re
 import warnings
 from functools import partial
 from pathlib import Path
+from platform import system
 from typing import Dict, List, Tuple, Union, Callable
 
 import torch
@@ -39,8 +39,8 @@ from ..embeddings_models import EmbeddingsModelFactory
 from ..errors import FastTextModelError
 from ..metrics import nll_loss, accuracy
 from ..network import ModelFactory
-from ..pre_processing import trailing_whitespace_cleaning, double_whitespaces_cleaning
 from ..pre_processing import coma_cleaning, lower_cleaning, hyphen_cleaning
+from ..pre_processing import trailing_whitespace_cleaning, double_whitespaces_cleaning
 from ..tools import CACHE_PATH, valid_poutyne_version
 from ..vectorizer import VectorizerFactory
 
@@ -1214,18 +1214,21 @@ class AddressParser:
                 )
 
     def _model_os_validation(self, num_workers):
-        if platform.system() == "Windows" and "fasttext" in self.model_type and num_workers > 0:
+        if system() == "Windows" and "fasttext" in self.model_type and num_workers > 0:
             raise FastTextModelError(
                 "On Windows system, we cannot use FastText-like models with parallelism workers since "
                 "FastText objects are not pickleable with the parallelism process use by Windows. "
                 "Thus, you need to set num_workers to 0 since 1 also means 'parallelism'."
             )
 
-        if platform.system().lower() == "darwin" and "fasttext" in self.model_type and num_workers > 0:
-            raise FastTextModelError(
+        if system() == "darwin" and "fasttext" in self.model_type and num_workers > 0:
+            torch.multiprocessing.set_start_method('fork')
+            warnings.warn(
                 "On MacOS system, we cannot use FastText-like models with parallelism out-of-the-box since "
                 "FastText objects are not pickleable with the parallelism process used by default by MacOS. "
-                "Thus, you need to set torch.multiprocessing.set_start_method('fork') to allow torch parallelism."
+                "Thus, we have set it to the 'fork' (i.e. torch.multiprocessing.set_start_method('fork'))"
+                " to allow torch parallelism.",
+                category=UserWarning,
             )
 
     def _apply_pre_processors(self, addresses: List[str]) -> List[str]: