+++ CHANGELOG.md
@@ -308,3 +308,6 @@
   by [Torch documentation](https://pytorch.org/tutorials/recipes/recipes/tuning_guide.html).
 - Add `torch.no_grad()` context manager in `__call__()` to increase performance.
 - Reduce memory swap between CPU and GPU by instantiating Tensor directly on the GPU device.
+- Improve some Warnings clarity (i.e. category and message).
+- Bug-fix MacOS multiprocessing. It was not possible to use in multiprocess, since we were not testing if torch
+  multiprocess was set properly. Now, we set it properly and raise a warning instead of an error.